Reverted the NetModule, PlayoneModule, and ApplicationComponent.
